requirements.txt
tweepy==3.10.0
Flask==2.0.1
flask-restful==0.3.9
Flask-SQLAlchemy==2.5.1
elasticsearch==7.13.0
python-logstash-async==2.3.0
prometheus-flask-exporter==0.18.2